Reflect form validation error in title
This adds guidance to prefix the `title` with something
(like the repeated error summary heading) to alert screen reader users
as soon as possible that there has been an error after submitting a form.
It also adds the required code to the two validation example forms.

This was very successfully user-tested with screen reader users,
with users actively pointing out how useful it is.