Change error summary ARIA role to "alert"
An ARIA `role="alert"` on the error summary should not be necessary
due to the focus on that area.
But without this VoiceOver on iOS doesn't announce the error message
at all due to this bug: https://bugs.webkit.org/show_bug.cgi?id=163719
Changing this doesn't change the behaviour in JAWS, NVDA or VO on macOS,
but it does mean that ZoomText is not announcing anything.
As VO on iOS is used much more often by blind people than ZoomText
(ZoomText is used more by visually impaired people who would see
the big red alert box), this seems to be a sensible sacrifice.